% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{A binomial distribution generator}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{Number of samples taken}

\item{n}{Sample size}

\item{p}{Probability of success}
}
\value{
A table of proportions by number of successes
}
\description{
A binomial distribution generator
}
\examples{
\dontrun{mybin(iter=50,n=10,p=0.6)}
}
