% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{title A CI for mean from a single sample}
\usage{
myci(x, conf.level = 0.95)
}
\arguments{
\item{x}{The reference sample}

\item{conf.level}{The desired confidence level}
}
\value{
A vector containing the lower and upper bounds of the CI
}
\description{
title A CI for mean from a single sample
}
\examples{
\dontrun set.seed(23);x = rnorm(30,mean=10,sd=12);myci(x)
}
