% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{title A Poisson variant of the Central Limit Theorem demonstration}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{The number of elements in each sample}

\item{iter}{the number of samples taken}

\item{lambda}{The lambda value of the Poisson distribution}

\item{...}{Optional arguments to improve the output}
}
\value{
Plots reflecting the distribution of sample statistics
}
\description{
title A Poisson variant of the Central Limit Theorem demonstration
}
\examples{
\dontrun mycltp(n=2,iter=10000,lambda=4)
}
