% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstfun.R
\name{myfirstfun}
\alias{myfirstfun}
\title{My first R function}
\usage{
myfirstfun(x)
}
\arguments{
\item{x}{A vector of quantitative data}
}
\value{
A vector of squared components
}
\description{
My first R function
}
\examples{
y <- 1:10; myfirstfun(y)
}
