% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{title A maximum likelihood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{The distribution function to sample from}

\item{x}{A vector of measurements}

\item{param}{A range of values to check for parameter}

\item{...}{Additional plot details}
}
\value{
A list of data about a plot, and a plot
}
\description{
title A maximum likelihood function
}
\examples{
\dontrun mymaxlik(function(x,param) log(dbinom(x,prob=param,size=20)),
\dontrun     c(1,4,2),seq(0,1,length=1000))
}
