% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{title A function that plots a normal curve with a lower-tail region}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean of the normal distribution}

\item{sigma}{The standard deviation of the normal distribution}

\item{a}{The upper bound of the probability}
}
\value{
A plot with a shaded region
}
\description{
title A function that plots a normal curve with a lower-tail region
}
\examples{
\dontrun{myncurve(mu=10,sigma=5,a=6)}
}
