% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplot.R
\name{myplot}
\alias{myplot}
\title{An introductory plot function}
\usage{
myplot(x)
}
\arguments{
\item{x}{A quantitative vector}
}
\value{
A plot of a downward parameter with the given domain
}
\description{
An introductory plot function
}
\examples{
\dontrun{d <- 1:40; myplot(x = d)}
}
