% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{function for p-values}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{a calculated value of t}

\item{xmax}{the maximum value of x to be plotted}

\item{n}{the sample size}

\item{alpha}{the value of alpha, e.g. 1-confidence level}
}
\value{
a plot with p-value areas, quantiles, and a p-value
}
\description{
function for p-values
}
\examples{
\dontrun mypvalue(t0=2.3563,n=30,alpha=0.05)
}
