% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myread.R
\name{myread}
\alias{myread}
\title{A .csv file reader}
\usage{
myread(csv)
}
\arguments{
\item{csv}{The filename to be read in}
}
\value{
A dataframe populated by the data from the .csv file
}
\description{
A .csv file reader
}
\examples{
\dontrun{file <- "SPRUCE.csv"; myread(file)}
}
